feat(frontend): make console the default landing page (#163)
- Merge console layout (bg + WorkspaceSettingsDialog) into root layout.tsx
- Create new app/page.tsx that renders AIConsole at "/"
- Delete marketing landing page ((marketing)/layout.tsx, page.tsx)
- Delete (console)/layout.tsx (merged into root)
- Convert /console to redirect "/" for backward compatibility
- /console/capabilities, /console/llm, /console/database still work
