+++ frontend/src/app/layout.tsx
@@ -4,6 +4,7 @@ import './globals.css'
 import { Providers } from './providers'
 import { GeistSans, GeistMono } from '@/lib/fonts'
 import { LOCALE_COOKIE_NAME, parseLocaleCookieValue } from '@/lib/locale-preference'
+import { WorkspaceSettingsDialog } from '@/components/settings/workspace-settings-dialog'
 
 // Locale-aware SSR reads the preference cookie during the request, so this layout is intentionally dynamic.
 export const dynamic = 'force-dynamic'
@@ -26,7 +27,12 @@ export default function RootLayout({
     <html lang={htmlLang} className={`${GeistSans.variable} ${GeistMono.variable}`} suppressHydrationWarning>
       <body suppressHydrationWarning>
         <Providers initialLocale={initialLocale}>
-          {children}
+          <div className="min-h-screen flex flex-col bg-slate-50/95 text-foreground xl:h-screen xl:overflow-hidden dark:bg-slate-950/95 dark:text-foreground">
+            <main className="w-full flex-1 min-h-0 xl:overflow-y-auto">
+              {children}
+            </main>
+            <WorkspaceSettingsDialog />
+          </div>
         </Providers>
       </body>
     </html>
