+++ frontend/src/app/page.tsx
@@ -1,0 +1,9 @@
+import { AIConsole } from '@/components/chat'
+
+export default function HomePage() {
+  return (
+    <div className="h-full">
+      <AIConsole />
+    </div>
+  )
+}
